+++ src/bot.js
@@ -1,15 +1,15 @@
 import { createSurvey } from "./lib/createSurvey.js";
 import { makeSurveyPost } from "./lib/makeSurveyPost.js";
 import { startAutoPosting } from "./lib/startAutoPosting.js";
-import { command } from "./utils/commandBuilder.js";
-import { handleAutoPost } from "./utils/handleAutoPost.js";
-import { handleCreate } from "./utils/handleCreate.js";
-import { handleCreateModal } from "./utils/handleModals.js";
-import { handleInfo } from "./utils/handleInfo.js";
-import { handleRespondModal } from "./utils/handleModals.js";
-import { handleRespond } from "./utils/handleRespond.js";
-import { handleRespondButton } from "./utils/handleRespondButton.js";
-import { handleView } from "./utils/handleView.js";
+import { command } from "./handlers/commandBuilder.js";
+import { handleAutoPost } from "./handlers/handleAutoPost.js";
+import { handleCreate } from "./handlers/handleCreate.js";
+import { handleCreateModal } from "./handlers/handleModals.js";
+import { handleInfo } from "./handlers/handleInfo.js";
+import { handleRespondModal } from "./handlers/handleModals.js";
+import { handleRespond } from "./handlers/handleRespond.js";
+import { handleRespondButton } from "./handlers/handleRespondButton.js";
+import { handleView } from "./handlers/handleView.js";
 
 import {
   maxResponsesForMultiResponsePerUser,
