+++ src/handlers/handleAutoPost.js
