+++ src/handlers/handleInfo.js
