+++ src/handlers/handleView.js
